Enum ohmers::OhmerError

pub enum OhmerError {
    NotSaved,
    RedisError(RedisError),
    EncoderError(EncoderError),
    DecoderError,
    UnknownIndex(String),
    UniqueIndexViolation(String),
    CommandError(Vec<u8>),
}

Variants

NotSaved

The operation requires the object to have an id, but it was never saved

RedisError

Error communicating with the server

EncoderError

Error encoding the object

DecoderError

Error decoding the object

UnknownIndex

A unique field has no value. The field name is returned.

UniqueIndexViolation

A unique field value is already in use. The field name is returned.

CommandError

There was an error translating a field to a string using utf8.