Struct ohmers::Collection

pub struct Collection<T: Ohmer> {
    // some fields omitted
}

A wrapper for classes that are referenced from another classes property.

Examples

model!(
    Tweeter {
        name:String = "".to_string();
        tweets:Collection<Tweet> = Collection::new();
    });
model!(
    Tweet {
        indices { tweeter:Reference<Tweeter> = Reference::new(); };
        description:String = "".to_string();
    });
let tweeter = create!(Tweeter { name: "Alice".to_owned(), }, &client).unwrap();
let mut tweet1 = create!(Tweet { tweeter: Reference::with_value(&tweeter), description: "foo".to_owned() }, &client).unwrap();
let mut tweet2 = create!(Tweet { tweeter: Reference::with_value(&tweeter), description: "bar".to_owned() }, &client).unwrap();
assert_eq!(collection!(tweeter.tweets, &client).sort("description", None, true, true).unwrap().collect::<Vec<_>>(),
        vec![
        tweet2,
        tweet1,
        ]
);

Methods

impl<T: Ohmer> Collection<T>

fn new() -> Self

fn all<'a, P: Ohmer>(&'a self, property: &str, parent: &P, r: &'a Client) -> Query<T>

Returns a query for all T elements referencing this object.