Struct ohmers::List

pub struct List<T: Ohmer> {
    // some fields omitted
}

A list of elements.

Examples

model!(
    Task {
        body:String = "".to_string();
    });
model!(
    Queue {
        tasks: List<Task> = List::new();
    });
let queue = create!(Queue {}, &client).unwrap();
let t1 = create!(Task { body: "task1".to_string() }, &client).unwrap();
let t2 = create!(Task { body: "task2".to_string() }, &client).unwrap();
let t3 = create!(Task { body: "task3".to_string() }, &client).unwrap();
push_back!(queue.tasks, t1, client).unwrap();
push_back!(queue.tasks, t2, client).unwrap();
push_back!(queue.tasks, t3, client).unwrap();
assert_eq!(len!(queue.tasks, client).unwrap(), 3);
assert_eq!(&*pop_front!(queue.tasks, client).unwrap().unwrap().body, "task1");
assert_eq!(len!(queue.tasks, client).unwrap(), 2);

Methods

impl<T: Ohmer> List<T>

fn new() -> Self

fn len<P: Ohmer>(&self, property: &str, parent: &P, r: &Client) -> Result<usize, OhmerError>

Number of items in the list.

fn push_back<P: Ohmer>(&self, property: &str, parent: &P, obj: &T, r: &Client) -> Result<(), OhmerError>

Adds an element at the end of the list.

fn pop_back<P: Ohmer>(&self, property: &str, parent: &P, r: &Client) -> Result<Option<T>, OhmerError>

Takes an element from the end of the list.

fn push_front<P: Ohmer>(&self, property: &str, parent: &P, obj: &T, r: &Client) -> Result<(), OhmerError>

Adds an element at the beginning of the list.

fn pop_front<P: Ohmer>(&self, property: &str, parent: &P, r: &Client) -> Result<Option<T>, OhmerError>

Takes an element from the beginning of the list.

fn first<P: Ohmer>(&self, property: &str, parent: &P, r: &Client) -> Result<Option<T>, OhmerError>

Retrieves an element from the beginning of the list.

fn last<P: Ohmer>(&self, property: &str, parent: &P, r: &Client) -> Result<Option<T>, OhmerError>

Retrieves an element from the end of the list.

fn try_range<'a, P: Ohmer>(&'a self, property: &str, parent: &P, start: isize, end: isize, r: &'a Client) -> Result<Iter<T>, OhmerError>

Creates an iterator for the list between start and end. Negative indices start from the end.

fn try_iter<'a, P: Ohmer>(&'a self, property: &str, parent: &P, r: &'a Client) -> Result<Iter<T>, OhmerError>

Creates an iterator for all the elements in the list.

fn contains<P: Ohmer>(&self, property: &str, parent: &P, obj: &T, r: &Client) -> Result<bool, OhmerError>

Checks if an element is in the list.

fn remove<P: Ohmer>(&self, property: &str, parent: &P, obj: &T, r: &Client) -> Result<usize, OhmerError>

Remove all occurrences of an element in the list.