Struct ohmers::Query

pub struct Query<'a, T: 'a + Ohmer> {
    // some fields omitted
}

A query of a set, or a result of set operations.

Examples

model!(
    derive { Clone }
    Dog {
        indices {
            age:u8 = 0;
            color:String = "".to_owned();
        };
        name:String = "".to_owned();
    });
create!(Dog { name: "Max".to_string(), age: 3, color: "white".to_string() }, &client).unwrap();
let buddy = create!(Dog { name: "Buddy".to_string(), age: 3, color: "black".to_string() }, &client).unwrap();
create!(Dog { name: "Bella".to_string(), age: 2, color: "black".to_string() }, &client).unwrap();
let lola = create!(Dog { name: "Lola".to_string(), age: 3, color: "black".to_string() }, &client).unwrap();
assert_eq!(Query::<Dog>::find("age", "3", &client).inter("color", "black").sort("name", None, true, true).unwrap().collect::<Vec<_>>(),
    vec![
        buddy.clone(),
        lola.clone(),
    ]);

Methods

impl<'a, T: Ohmer> Query<'a, T>

fn new(set: Set, r: &'a Client) -> Self

Create a new Query for a Set

fn from_keys(kv: &[(&str, &str)], r: &'a Client) -> Self

Creates a new query with the intersection of all key/value

fn key(field: &str, value: &str) -> Set

Creates the stal set for a key/value combination

fn find(field: &str, value: &str, r: &'a Client) -> Self

Creates a query for a key/value combination

fn inter(&mut self, field: &str, value: &str) -> &mut Self

Updates the set to be the intersection of the current one and the set where field=value.

fn sinter(&mut self, sets: Vec<Set>)

Updates the set to be the intersection of the current set and all given sets.

fn union(&mut self, field: &str, value: &str) -> &mut Self

Updates the set to be the union of the current one and the set where field=value.

fn sunion(&mut self, sets: Vec<Set>)

Updates the set to be the union of the current set and all given sets.

fn diff(&mut self, field: &str, value: &str) -> &mut Self

Updates the set to remove all elements where field=value.

fn sdiff(&mut self, sets: Vec<Set>)

Updates the set to remove all elements in any of the provided sets.

fn try_iter(&self) -> Result<Iter<'a, T>, OhmerError>

Creates an iterator for all objects in the set.

fn try_into_iter(self) -> Result<Iter<'a, T>, OhmerError>

Creates an iterator for all objects in the set, consuming the query.

fn sort(&self, by: &str, limit: Option<(usize, usize)>, asc: bool, alpha: bool) -> Result<Iter<'a, T>, OhmerError>

Creates an iterator for all objects in the set sorted by by.